added more docs, moving client to *unix only module
