+++ cmd/agent/README.md
@@ -19,23 +19,15 @@ DD_API_KEY=12345678990 ./bin/agent/agent
 Once the Agent has started, you can interact with it using the `agent` binary, 
 for example to activate the shutdown process just type `agent stop`.
 
-The agent communicates with the outside world either locally with standard IPC
-strategies or through an HTTP API to ease the development of 3rd party tools
-and interfaces.
+The agent communicates with the outside world through an HTTP API to ease the 
+development of 3rd party tools and interfaces. Since HTTP is transported over
+a [Unix Socket][0] on *nix platform and [Named Pipes][1] on Windows, authorization
+is delegated to the filesystem.
 
-### HTTP API
-Still work in progress.
-
-### Local IPC
-There are a set of operations that cannot be performed through an HTTP API for
-security reasons, like asking the agent to stop or retrieve auth tokens to use 
-with the HTTP API. In such cases a [Unix Socket][0] is used on *nix platform 
-while [Named Pipes][1] are used on Windows (still WIP). Commands are simple
-strings, for the time being clients should not expect anything back from the
-server.
-
-The commands accepted through the IPC interface:
- * `stop` to trigger the shutdown procedure
+Endpoints implemented so far (this list should be killed in favor of [swagger][2] at some point):
+    * [GET] http://localhost/agent/version
+    * [POST] http://localhost/agent/stop
 
 [0]: https://en.wikipedia.org/wiki/Unix_domain_socket
 [1]: https://msdn.microsoft.com/en-us/library/windows/desktop/aa365590.aspx
+[2]: http://swagger.io/
